# perf_script_bug.py:perf probe is failing