--- a/perf/perf_script_bug.py.data/probe.cfg
+++ b/perf/perf_script_bug.py.data/probe.cfg
@@ -1,0 +1,4 @@
+[rhel]
+probepoint = 3
+[SuSE]
+probepoint = 4