changed the probe command according to that

Signed-off-by: shirisha Ganta <shiganta@in.ibm.com>