@shirishaganta 2020-07-21
Before Patch:
root@ltczz256-lp3 perf]# avocado run perf_script_bug.py
/usr/local/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : 3379f6dce4aabd3dd5b70ad67d78897538b3a541
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-07-21T00.19-3379f6d/job.log
 (1/1) perf_script_bug.py:PerfProbe.test_probe: ERROR: Command 'perf probe -x perf_test 'perf_test.c:4'' failed.\nstdout: b''\nstderr: b'This line is sharing the address with other lines.\nPlease try to probe at /var/tmp/avocado_15hz5ie0/perf_test.c:3 instead.\n  Error: Failed to add events.\n'\nadditional_info:... (1.10 s)
RESULTS    : PASS 0 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-07-21T00.19-3379f6d/results.html
JOB TIME   : 1.20 s
[root@ltczz256-lp3 perf]# 
After Patch:
[root@ltczz256-lp3 perf]# vi perf_script_bug.py
[root@ltczz256-lp3 perf]# avocado run perf_script_bug.py
/usr/local/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : 4edca24148eecaa01dd0c93eb2389775061202c4
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-07-21T00.45-4edca24/job.log
 (1/1) perf_script_bug.py:PerfProbe.test_probe: PASS (2.73 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-07-21T00.45-4edca24/results.html
JOB TIME   : 2.97 s
[root@ltczz256-lp3 perf]#