@shirishaganta 2020-07-23
on python2 old code worked. But in python 3 it is not taking that position "4". Now I checked in both rhel8.2 and rhel8.3 with insert position 3 it is working fine