@shirishaganta 2020-07-27
Earlier I mentioned the changes on python 2 and python 3. Later I got to know those are because of  recent Rhel release, many perf related changes went in  and now because of this changes we changed the insertion point to perf probe -x perf_test 'perf_test.c:3' in order to work successfully