@shirishaganta 2020-07-30
Tested the above patch on sles15sp2, Rhel8.2 and Rhel8.3. It is working fine
sles15sp2 After Patch:
localhost:/home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf # avocado run perf_script_bug.py
/usr/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : bdb189f998697cb650eb6d2b36d857ae1f4222c1
JOB LOG    : /root/avocado-fvt-wrapper/results/job-2020-07-30T12.43-bdb189f/job.log
 (1/1) perf_script_bug.py:PerfProbe.test_probe: PASS (2.31 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /root/avocado-fvt-wrapper/results/job-2020-07-30T12.43-bdb189f/results.html
JOB TIME   : 2.40 s
localhost:/home/avocado-fvt-wrapper/tests/avocado-misc-tests/perf #

Rhel8.3 After Patch:
[root@zzfp219-lp4 perf]# avocado run perf_script_bug.py
/usr/local/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : 802e30752f6b3c5774fd46219483c23c97482b81
JOB LOG    : /home/OpTest/avocado-fvt-wrapper/results/job-2020-07-30T03.37-802e307/job.log
 (1/1) perf_script_bug.py:PerfProbe.test_probe: PASS (13.46 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/OpTest/avocado-fvt-wrapper/results/job-2020-07-30T03.37-802e307/results.html
JOB TIME   : 14.38 s
[root@zzfp219-lp4 perf]#