@PraveenPenguin 2020-07-30
> This pull request **fixes 2 alerts** when merging [b59d672](https://github.com/avocado-framework-tests/avocado-misc-tests/commit/b59d67209e2689a0bff40366771bacd95093af25) into [4b5beba](https://github.com/avocado-framework-tests/avocado-misc-tests/commit/4b5beba67df9dc3b872043e03bf9e4bcaf7fea88) - [view on LGTM.com](https://lgtm.com/projects/g/avocado-framework-tests/avocado-misc-tests/rev/pr-f5af74bd8c47ad8ba69dd9197b4be8461becbea1)
> 
> **fixed alerts:**
> 
>     * 2 for Variable defined multiple times

@shirishaganta why you pushing the change it is valid alert so we are good on this no need to resolve this