@shirishaganta 2020-10-09
After Patch Rhel:
[root@ltcfleet8-lp2 perf]# avocado run perf_script_bug.py
JOB ID     : 00746f0e9f2095bede7c1cd992b11b1777565c6b
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-10-08T05.47-00746f0/job.log
 (1/1) perf_script_bug.py:PerfProbe.test_probe: PASS (3.19 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-10-08T05.47-00746f0/results.html
JOB TIME   : 3.72 s
[root@ltcfleet8-lp2 perf]#
After Patch Sles:
localhost:/home/sirisha/avocado-fvt-wrapper/tests/avocado-misc-tests/perf # 
localhost:/home/sirisha/avocado-fvt-wrapper/tests/avocado-misc-tests/perf # avocado run perf_script_bug.py
JOB ID     : 7b15650321b8d055aaa42284488141ae6f6c9944
JOB LOG    : /home/sirisha/avocado-fvt-wrapper/results/job-2020-10-07T23.46-7b15650/job.log
 (1/1) perf_script_bug.py:PerfProbe.test_probe: PASS (2.22 s)
RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/sirisha/avocado-fvt-wrapper/results/job-2020-10-07T23.46-7b15650/results.html
JOB TIME   : 2.81 s
localhost:/home/sirisha/avocado-fvt-wrapper/tests/avocado-misc-tests/perf #